test(discord): cr-r8 nits for chaos validation (PR 15)
Four cr-r8 cleanups (verdict: "Nothing here is blocking"):

- Hoist `tableNamesTargeted` + `assertNoUnexpectedTableCalls`
  (plus the FLOW_STATE_TABLE_NAME import) from
  deploy-during-flow into helpers/chaos-ddb.js. Call the helper
  at the end of both resume-fail tests too — the watchdog's
  exhaustion-exit path is also gateway-tier and must not touch
  flow_state. One helper, two call sites, same regression
  surface covered uniformly.
- Lift `MAX_YIELDS_FOR_TRANSFER` and `MAX_YIELDS_FOR_PUSH_RESOLVE`
  into a shared `MAX_MICROTASK_YIELDS = 50` in chaos-ddb.js.
  Same bound, same purpose, one source of truth.
- Normalize the resume-fail test #2 heartbeat-row assertion to
  the same map+not.toContain shape used in test #1, so the two
  read identically.
- Move the spy-brittleness pointer from the file header to a
  comment immediately above the `jest.spyOn(leader, 'pushHandoff')`
  call. The warning lives where a future editor of those three
  lines actually sees it.

All 2408 tests pass.
